// wardn/frontend/lib/api/generated/model/mCPServerCreate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { _MCPServerCreateMeta } from './_mCPServerCreateMeta';
import type { MCPServerCreateIconsItem } from './mCPServerCreateIconsItem';
import type { MCPServerCreatePackagesItem } from './mCPServerCreatePackagesItem';
import type { MCPServerCreateRemotesItem } from './mCPServerCreateRemotesItem';
import type { MCPServerCreateRepository } from './mCPServerCreateRepository';

export interface MCPServerCreate {
  /** @minLength 1 */
  $schema: string;
  _meta?: _MCPServerCreateMeta;
  /** @minLength 1 */
  description: string;
  icons?: MCPServerCreateIconsItem[];
  /**
     * @minLength 3
     * @maxLength 200
     * @pattern ^[a-zA-Z0-9.-]+/[a-zA-Z0-9._-]+$
     */
  name: string;
  packages?: MCPServerCreatePackagesItem[];
  remotes?: MCPServerCreateRemotesItem[];
  repository?: MCPServerCreateRepository;
  /** @maxLength 100 */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  version: string;
  /** @maxLength 2048 */
  websiteUrl?: string;
  [key: string]: unknown;
 }